ДЗ 5го урока, changes
