+++ Lesson5/Problem_4.txt
@@ -1,0 +1,4 @@
+One - 1
+Two - 2
+Three - 3
+Four - 4
