+++ Lesson5/Problem_6.py
@@ -2,28 +2,11 @@
 # Важно, чтобы для каждого предмета не обязательно были все типы занятий.
 # Сформировать словарь, содержащий название предмета и общее количество занятий по нему. Вывести словарь на экран.
 
+from my_func import didgits_of_string
+
 # Создаем список из строк файла
-with open('Lesson6.txt', 'r') as f_objects:
+with open('Problem_6.txt', 'r') as f_objects:
     l_objects = f_objects.readlines()
-
-
-def didgits_of_string(i_string: str):
-    """Функция возвращает сумму чисел из исходной строки
-       Корректно работает если в каждом слове строки только одно число
-    """
-    words = i_string.split()  # Разобьем входную строку на слова
-    digit_list = []  # Выходной Список чисел
-    for word in words:
-        s_didgit = ''
-        for symbol in word:
-            if symbol.isdigit():
-                s_didgit += symbol
-        try:
-            digit_list.append(int(s_didgit))
-        except:
-            pass
-    return digit_list
-
 
 d_objects = {}  # Наш словарь предметов
 for object in l_objects:
