+++ Lesson5/Problem_6.txt
@@ -1,0 +1,5 @@
+Информатика: 100(л) 50(пр) 20(лаб)
+Физика: 30(л) - 10(лаб)
+Физкультура: - 30(пр) -
+Философия: 34(л) - -
+Конная езда: 10(л) 40(пр) -
